+++ R/utils_nanonext.R
@@ -67,10 +67,3 @@ listener_connected <- function(dialer) {
   connection_opened(dialer) &&
     nanonext::stat(dialer$dialer[[1]], "pipes") > 0L
 }
-
-# TODO: just use .unresolved() when nanonext > 0.8.0.9000 reaches CRAN.
-nanonext_unresolved <- if_any(
-  exists(x = ".unresolved", envir = getNamespace(name = "nanonext")),
-  eval(expr = parse(text = "nanonext::.unresolved")),
-  nanonext::unresolved
-)